% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_skeleton.R
\name{adj_m_to_centrality}
\alias{adj_m_to_centrality}
\title{Calculate centrality for each node based on the adjacency matrix}
\usage{
adj_m_to_centrality(
  input_adj_m,
  type = c("degree", "betweenness", "closeness")
)
}
\description{
Calculate centrality for each node based on the adjacency matrix
}
