% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_skeleton.R
\name{bringmann_2017_dataset2}
\alias{bringmann_2017_dataset2}
\title{Creates empirical graph of Bringmann's paper
Creates a 10-node network based on Bringmann's group networks for Dataset-2
doi:10.1177/1073191116645909}
\usage{
bringmann_2017_dataset2(abs_cst = 0.01, eig_cst = 0.5, drop_index = NULL)
}
\description{
Creates empirical graph of Bringmann's paper
Creates a 10-node network based on Bringmann's group networks for Dataset-2
doi:10.1177/1073191116645909
}
