% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_skeleton.R
\name{clamp_min_weights}
\alias{clamp_min_weights}
\title{Helper function to ensure the weights are at least a certain magnitude}
\usage{
clamp_min_weights(weights, abs_cst)
}
\description{
Helper function to ensure the weights are at least a certain magnitude
}
