% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fitted_results.R
\name{collate_results_timelen}
\alias{collate_results_timelen}
\title{Function to collate the VAR results for different time lengths}
\usage{
collate_results_timelen(fitted_list, true_Phi, true_Psi)
}
\description{
Function to collate the VAR results for different time lengths
}
