% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fitted_results_omitvar.R
\name{collate_results_timelen_omitvar}
\alias{collate_results_timelen_omitvar}
\title{Function to collate the VAR results for different time lengths}
\usage{
collate_results_timelen_omitvar(
  fitted_list,
  true_Phi,
  true_Psi,
  omitted_var_index
)
}
\description{
Function to collate the VAR results for different time lengths
}
