% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fitted_results_prepost.R
\name{collate_results_timelen_prepost}
\alias{collate_results_timelen_prepost}
\title{Function to collate the VAR results for different time lengths}
\usage{
collate_results_timelen_prepost(
  pre_fitted_list,
  post_fitted_list,
  pre_true_Phi,
  post_true_Phi
)
}
\description{
Function to collate the VAR results for different time lengths
}
