% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fitted_results_prepost.R
\name{compare_prepost}
\alias{compare_prepost}
\title{Function to compare prepost}
\usage{
compare_prepost(pre_lav_fitobj, post_lav_fitobj, pre_true_Phi, post_true_Phi)
}
\description{
Function to compare prepost
}
