% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fitted_results.R
\name{compare_true_with_est}
\alias{compare_true_with_est}
\title{Function to compare estimated with true parameters       ( THIS IS FOR ONE TIME LENGTH ONLY; NEED TO LAPPLY )}
\usage{
compare_true_with_est(lav_fitobj, true_Phi, true_Psi)
}
\description{
Function to compare estimated with true parameters       ( THIS IS FOR ONE TIME LENGTH ONLY; NEED TO LAPPLY )
}
