% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fitted_results_omitvar.R
\name{compare_true_with_est_omitvar}
\alias{compare_true_with_est_omitvar}
\title{Function to compare estimated with true parameters       ( THIS IS FOR ONE TIME LENGTH ONLY; NEED TO LAPPLY )}
\usage{
compare_true_with_est_omitvar(
  lav_fitobj,
  true_Phi,
  true_Psi,
  omitted_var_index
)
}
\description{
Function to compare estimated with true parameters       ( THIS IS FOR ONE TIME LENGTH ONLY; NEED TO LAPPLY )
}
