% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fitted_results_prepost.R
\name{compute_prepost_gramian_metrics}
\alias{compute_prepost_gramian_metrics}
\title{Compute pre-post metrics for gramian
Convenience function to extract max node(s)}
\usage{
compute_prepost_gramian_metrics(pre_gramian, post_gramian)
}
\description{
Compute pre-post metrics for gramian
Convenience function to extract max node(s)
}
