% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fitted_results.R
\name{compute_VAR_metrics}
\alias{compute_VAR_metrics}
\title{Function to compute metrics of interest given lavaan fit and true values}
\usage{
compute_VAR_metrics(true_params, est_params)
}
\description{
Function to compute metrics of interest given lavaan fit and true values
}
