% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fitted_results_omitvar.R
\name{compute_VAR_metrics_omitvar}
\alias{compute_VAR_metrics_omitvar}
\title{Function to compute metrics of interest given lavaan fit and true values}
\usage{
compute_VAR_metrics_omitvar(true_params, est_params, omitted_var_index)
}
\description{
Function to compute metrics of interest given lavaan fit and true values
}
