% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_simulation_data.R
\name{create_sim_obj}
\alias{create_sim_obj}
\title{Creates simulation object based on simulation parameters}
\usage{
create_sim_obj(phi.func, phi.func.args = list(), num.ord.max = 7)
}
\arguments{
\item{num.ord.max}{}
}
\description{
Creates simulation object based on simulation parameters
}
