% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fitted_results.R
\name{extract_mc_results}
\alias{extract_mc_results}
\title{We need a function here that will extract the info from a list of (list of) fits
Outer list is for MC iterations
Inner list is for truncated timepoints}
\usage{
extract_mc_results(mc_fitted_list, true_Phi, true_Psi)
}
\description{
We need a function here that will extract the info from a list of (list of) fits
Outer list is for MC iterations
Inner list is for truncated timepoints
}
