% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fitted_results_omitvar.R
\name{extract_mc_results_omitvar}
\alias{extract_mc_results_omitvar}
\title{We need a function here that will extract the info from a list of (list of) fits
Outer list is for MC iterations
Inner list is for truncated timepoints}
\usage{
extract_mc_results_omitvar(
  mc_fitted_list,
  true_Phi,
  true_Psi,
  omitted_var_index
)
}
\description{
We need a function here that will extract the info from a list of (list of) fits
Outer list is for MC iterations
Inner list is for truncated timepoints
}
