% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fitted_results_prepost.R
\name{extract_mc_results_prepost}
\alias{extract_mc_results_prepost}
\title{We need a function here that will extract the info from a list of (list of) fits
Outer list is for MC iterations
Inner list is for truncated timepoints}
\usage{
extract_mc_results_prepost(
  mc_pre_fitted_list,
  mc_post_fitted_list,
  pre_true_Phi,
  post_true_Phi
)
}
\description{
We need a function here that will extract the info from a list of (list of) fits
Outer list is for MC iterations
Inner list is for truncated timepoints
}
