% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaan_VAR.R
\name{fit_lavaan_VAR}
\alias{fit_lavaan_VAR}
\title{Wrapper to fit a VAR in lavaan using the block-toeplitz method}
\usage{
fit_lavaan_VAR(
  input_ts,
  force_reorder = FALSE,
  ord_as_cont = FALSE,
  estimator = "DEFAULT"
)
}
\description{
Wrapper to fit a VAR in lavaan using the block-toeplitz method
}
