% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaan_VAR.R
\name{gen_lavaan_dfm_syntax}
\alias{gen_lavaan_dfm_syntax}
\title{Generates the lavaan syntax for a given VAR}
\usage{
gen_lavaan_dfm_syntax(
  num_vars,
  var_thres = NULL,
  omit_thres = NULL,
  ordinal = TRUE
)
}
\arguments{
\item{num_vars}{number of variables in the system}

\item{var_thres}{named list of thresholds for each of the ordinal variables}

\item{ordinal}{boolean indicating if the variables are ordinal}
}
\value{
a string in lavaan format for model fitting
}
\description{
Generates the lavaan syntax for a given VAR
}
