% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_simulation_data.R
\name{generate_TS_from_simobj}
\alias{generate_TS_from_simobj}
\title{Generate multiple TS from a given simulation object}
\usage{
generate_TS_from_simobj(sim_obj, num_mc_samples, max_timepts)
}
\description{
Generate multiple TS from a given simulation object
}
