% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_skeleton.R
\name{random_Psi}
\alias{random_Psi}
\title{Need a function to generate a random positive definite Psi matrix}
\usage{
random_Psi(nvars)
}
\description{
Need a function to generate a random positive definite Psi matrix
}
