% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_skeleton.R
\name{rank_nodes_centrality}
\alias{rank_nodes_centrality}
\title{Rank order the nodes based on centrality}
\usage{
rank_nodes_centrality(
  input_adj_m,
  type = c("degree", "betweenness", "closeness")
)
}
\description{
Rank order the nodes based on centrality
}
