% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_skeleton.R
\name{run_and_plot_simulation}
\alias{run_and_plot_simulation}
\title{Runs and plots the results from the simulation}
\usage{
run_and_plot_simulation(
  m_func,
  m_func_args = list(),
  num_iter = 1000,
  centrality_type = c("degree", "betweenness", "closeness"),
  metric_type = c("max_node", "dgramian"),
  plot_type = c("boxplot", "histogram", "correlation")
)
}
\description{
Runs and plots the results from the simulation
}
