% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models_simulation_data.R
\name{run_simulation_multi_ts}
\alias{run_simulation_multi_ts}
\title{Function to mclapply over the different time series iterations}
\usage{
run_simulation_multi_ts(
  input_ts_list,
  save_dir = NULL,
  mc.cores = 4,
  timepts_vec = c(50, 100, 200, 500, 1000),
  ord_as_cont = FALSE,
  estimator = "DEFAULT"
)
}
\arguments{
\item{input_ts_list}{list of time-series from the same VAR model}

\item{save_dir}{location to save the fitted models}

\item{ord_as_cont}{force ordinal variables to be continuous}

\item{estimator}{string indicating which estimator to use}
}
\description{
Function to mclapply over the different time series iterations
}
