% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models_simulation_data.R
\name{run_simulation_one_ts}
\alias{run_simulation_one_ts}
\title{Function to run simulation for one timeseries}
\usage{
run_simulation_one_ts(
  input_ts,
  timepts_vec = c(50, 100, 200, 500, 1000),
  ord_as_cont = FALSE,
  estimator = "DEFAULT"
)
}
\arguments{
\item{input_ts}{a dataframe of dimension (ntimepts x numvars)}

\item{timepts_vec}{vector of timepoints used for the simulations}

\item{ord_as_cont}{force ordinal variables to be continuous}

\item{estimator}{string indicating which estimator to use}
}
\description{
Function to run simulation for one timeseries
}
