% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_skeleton.R
\name{scale_free_seeded}
\alias{scale_free_seeded}
\title{Creates fixed scale-free graph}
\usage{
scale_free_seeded(
  num_nodes = 10,
  abs_cst = 0.01,
  eig_cst = 0.5,
  prop_zero = 0.8,
  prob_pos = 0.5,
  mode_in = c("acyclic", "random"),
  seed = 100,
  generate_skeleton = FALSE
)
}
\description{
Creates fixed scale-free graph
}
