% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_skeleton.R
\name{sim_multiple_scale_free}
\alias{sim_multiple_scale_free}
\title{Runs and plots simulation from across different scale_free structures}
\usage{
sim_multiple_scale_free(num_iter = 1000, num_seeds = 100)
}
\description{
Runs and plots simulation from across different scale_free structures
}
