% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_skeleton.R
\name{simulate_gramian}
\alias{simulate_gramian}
\title{Compute gramian for random parameter values}
\usage{
simulate_gramian(m_func, num_iter = 1000, m_func_args = list())
}
\description{
Compute gramian for random parameter values
}
