% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateVAR.R
\name{simulateVAR}
\alias{simulateVAR}
\title{R6 class for simulating VARs
Should add a check that indicates if the given Phi and Psi results in standardized TS}
\description{
R6 class for simulating VARs
Should add a check that indicates if the given Phi and Psi results in standardized TS

R6 class for simulating VARs
Should add a check that indicates if the given Phi and Psi results in standardized TS
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-simulateVAR-new}{\code{simulateVAR$new()}}
\item \href{#method-simulateVAR-set_custom_name}{\code{simulateVAR$set_custom_name()}}
\item \href{#method-simulateVAR-set_custom_tags}{\code{simulateVAR$set_custom_tags()}}
\item \href{#method-simulateVAR-clear_saved_params}{\code{simulateVAR$clear_saved_params()}}
\item \href{#method-simulateVAR-initialize_models}{\code{simulateVAR$initialize_models()}}
\item \href{#method-simulateVAR-generate_ts_from_model}{\code{simulateVAR$generate_ts_from_model()}}
\item \href{#method-simulateVAR-clone}{\code{simulateVAR$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-simulateVAR-new"></a>}}
\if{latex}{\out{\hypertarget{method-simulateVAR-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{simulateVAR$new(
  Phi,
  Psi,
  num.ord.max = 7,
  burn.in = 1000,
  max.tries = 100,
  fixed_params = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-simulateVAR-set_custom_name"></a>}}
\if{latex}{\out{\hypertarget{method-simulateVAR-set_custom_name}{}}}
\subsection{Method \code{set_custom_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{simulateVAR$set_custom_name(modname)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-simulateVAR-set_custom_tags"></a>}}
\if{latex}{\out{\hypertarget{method-simulateVAR-set_custom_tags}{}}}
\subsection{Method \code{set_custom_tags()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{simulateVAR$set_custom_tags(tags)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-simulateVAR-clear_saved_params"></a>}}
\if{latex}{\out{\hypertarget{method-simulateVAR-clear_saved_params}{}}}
\subsection{Method \code{clear_saved_params()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{simulateVAR$clear_saved_params()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-simulateVAR-initialize_models"></a>}}
\if{latex}{\out{\hypertarget{method-simulateVAR-initialize_models}{}}}
\subsection{Method \code{initialize_models()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{simulateVAR$initialize_models()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-simulateVAR-generate_ts_from_model"></a>}}
\if{latex}{\out{\hypertarget{method-simulateVAR-generate_ts_from_model}{}}}
\subsection{Method \code{generate_ts_from_model()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{simulateVAR$generate_ts_from_model(
  time_len = 100,
  num.ord.out = NULL,
  initialize = FALSE,
  verbose = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-simulateVAR-clone"></a>}}
\if{latex}{\out{\hypertarget{method-simulateVAR-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{simulateVAR$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
