% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_VAR.R
\name{standardize_VAR}
\alias{standardize_VAR}
\title{Creates a standardized VAR from a given stationary VAR model}
\usage{
standardize_VAR(Phi_in, Psi_in)
}
\description{
Convenience function to check if Phi is stationary
Convenience function to check if Psi is PD
}
