% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_skeleton.R
\name{two_broken_chains}
\alias{two_broken_chains}
\title{Creates two broken chains with random parameters}
\usage{
two_broken_chains(abs_cst = 0.01, eig_cst = 0.5)
}
\description{
Creates two broken chains with random parameters
}
